# 🚀 GPU Acceleration: 10-100x Performance Boost for Large-Scale Name Matching