--- a/config-gpu.ini
+++ b/config-gpu.ini
@@ -1,0 +1,59 @@
+# GPU Acceleration Configuration for Name Matching
+# Copy this to config.ini and modify as needed
+
+[database]
+host = localhost
+port = 3306
+user = root
+password = 
+database = name_matching
+pool_size = 5
+max_overflow = 10
+pool_timeout = 30
+pool_recycle = 3600
+
+[gpu]
+# Enable or disable GPU acceleration
+enabled = true
+
+# Preferred GPU framework (auto, cupy, torch, numba)
+# auto = automatically select best available framework
+# cupy = Use CuPy (recommended for best performance)
+# torch = Use PyTorch
+# numba = Use Numba CUDA
+framework = auto
+
+# GPU device ID to use (0 for first GPU, 1 for second, etc.)
+device_id = 0
+
+# Batch size for GPU processing
+# Larger values use more GPU memory but may be faster
+# Adjust based on your GPU memory capacity
+batch_size = 1000
+
+# GPU memory limit in GB
+# The system will try to stay under this limit
+memory_limit_gb = 4.0
+
+# Fallback threshold - use CPU for datasets smaller than this
+# This avoids GPU overhead for small datasets
+fallback_threshold = 10000
+
+[matching]
+# Matching thresholds
+match_threshold = 0.75
+non_match_threshold = 0.55
+
+# Name component weights
+first_name_weight = 0.4
+middle_name_weight = 0.2
+last_name_weight = 0.3
+full_name_sorted_weight = 0.1
+
+# Additional field weights
+birthdate_weight = 0.3
+geography_weight = 0.3
+
+[logging]
+level = INFO
+format = %(asctime)s - %(name)s - %(levelname)s - %(message)s