--- a/requirements-gpu.txt
+++ b/requirements-gpu.txt
@@ -1,0 +1,34 @@
+# GPU Acceleration Requirements for Name Matching
+# Install these packages for GPU acceleration support
+
+# Core GPU frameworks (install at least one)
+# CuPy - Recommended for best performance
+cupy-cuda11x>=11.0.0  # For CUDA 11.x
+# cupy-cuda12x>=12.0.0  # For CUDA 12.x (uncomment if using CUDA 12)
+
+# PyTorch - Alternative GPU framework
+torch>=1.12.0
+torchvision>=0.13.0
+torchaudio>=0.12.0
+
+# Numba CUDA - Already included in main requirements
+# numba>=0.56.0  # Already in requirements.txt
+
+# Additional GPU utilities
+nvidia-ml-py3>=7.352.0  # For GPU monitoring
+pynvml>=11.4.1  # Alternative GPU monitoring
+
+# Memory management
+psutil>=5.8.0  # For system memory monitoring
+
+# Performance profiling
+line_profiler>=3.5.1  # For performance analysis
+memory_profiler>=0.60.0  # For memory usage analysis
+
+# Optional: Machine learning frameworks for advanced similarity models
+# scikit-learn>=1.1.0  # Already in main requirements
+# tensorflow>=2.9.0  # Uncomment for TensorFlow GPU support
+
+# Development and testing
+pytest-benchmark>=3.4.1  # For performance benchmarking
+pytest-xdist>=2.5.0  # For parallel testing